protobuf-kafka-cli/requirements.txt
protobuf
kafka-python
